Lint fixes, separate unit/e2e/integration tests, review fixes
