+++ .github/workflows/tests.yml
@@ -75,4 +75,4 @@ jobs:
 
       - name: Run Go tests
         run: |
-          make go-test
+          make test
