+++ sdk_tests/input_table_not_found.json
@@ -1,5 +1,5 @@
 {
   "describe_table": [
-    "input6"
+    "non_existent_table"
   ]
 }
